* Merged with mgdev a27116740a7e754b0113a010e7b103940549385e

+++ src/game/bg_public.h
@@ -915,7 +915,7 @@ typedef struct
 
   int       health;
   float     fallDamage;
-  int       regenRate;
+  float     regenRate;
 
   int       abilities;
